♻️ refactor:  determinePosition()を状態に変更
#74
